refactor(_DoLacnaClient): move dolacna_client to separate file and add categories to prompt in processors.py
